check not only if there is cookie but exactly session cookie

+++ paths/logout.js
@@ -3,7 +3,10 @@ const pool = require("../pool");
 const logout = async (req, res) => {
     try {
         const headers = req.headers
-        if(headers.cookie) {
+        if(
+            headers.cookie?.includes('session')
+            && headers?.cookie[ headers?.cookie.indexOf('session')+7 ] === '='
+        ) {
             const sessionId = headers.cookie.split('=')[1]
             await pool?.query(
                 'DELETE FROM sessions WHERE id = $1',
